+++ apps/dispatch-api/src/simulation-run/simulation-run.service.ts
@@ -7,6 +7,7 @@
 import {
   HttpService,
   Injectable,
+  Inject,
   InternalServerErrorException,
   Logger,
   NotFoundException,
@@ -28,7 +29,12 @@ import {
 } from '@biosimulations/dispatch/api-models';
 import { SimulationRunStatus } from '@biosimulations/datamodel/common';
 import { SharedStorageService } from '@biosimulations/shared/storage';
-
+import {
+  DispatchFailedPayload,
+  DispatchMessage,
+  DispatchProcessedPayload,
+} from '@biosimulations/messages/messages';
+import { ClientProxy } from '@nestjs/microservices';
 // 1gb in bytes to be used as file size limits
 const ONE_GIGABYTE = 1000000000;
 const toApi = <T extends SimulationRunModelType>(
@@ -49,14 +55,16 @@ export class SimulationRunService {
     private simulationRunModel: Model<SimulationRunModel>,
     private storageService: SharedStorageService,
     private http: HttpService,
+    @Inject('NATS_CLIENT') private client: ClientProxy,
   ) {}
 
   public async setStatus(
     id: string,
     status: SimulationRunStatus,
+    statusReason: string,
   ): Promise<SimulationRunModel | null> {
     const model = await this.getModel(id);
-    return this.updateModelStatus(model, status);
+    return this.updateModelStatus(model, status, statusReason);
   }
 
   /**
@@ -145,7 +153,7 @@ export class SimulationRunService {
 
     this.updateModelPublic(model, run.public);
     this.updateModelResultSize(model, run.resultsSize);
-    this.updateModelStatus(model, run.status);
+    this.updateModelStatus(model, run.status, run.statusReason);
 
     return toApi(await model.save());
   }
@@ -210,7 +218,7 @@ export class SimulationRunService {
     const url = body.url;
     // If the url provides the following information, grab it and store it in the database
     //! This does not adress the security issues of downloading user provided urls.
-    //! The content size may not be present or accurate. The backend must check the size. See @2536
+    //! The content size may not be present or accurate. The backend must check the size. See #2536
     let size = undefined;
     let mimetype = undefined;
     let originalname = undefined;
@@ -275,6 +283,7 @@ export class SimulationRunService {
   private updateModelStatus(
     model: SimulationRunModel | null,
     status: SimulationRunStatus | undefined,
+    statusReason: string | undefined,
   ): SimulationRunModel | null {
     let allowUpdate = true;
 
@@ -305,15 +314,18 @@ export class SimulationRunService {
     // Careful not to add else here
     if (model && status && allowUpdate) {
       model.status = status;
+      model.statusReason = statusReason;
       model.refreshCount = model.refreshCount + 1;
       this.logger.log(
         `Set ${model.id} status to ${model.status} on update ${model.refreshCount} `,
       );
-      if (
-        status == SimulationRunStatus.SUCCEEDED ||
-        status == SimulationRunStatus.FAILED
-      ) {
-        this.updateModelRunTime(model);
+      this.updateModelRunTime(model);
+      if (status == SimulationRunStatus.SUCCEEDED) {
+        const message = new DispatchFailedPayload(model.id);
+        this.client.emit(DispatchMessage.failed, message);
+      } else if (status == SimulationRunStatus.FAILED) {
+        const message = new DispatchProcessedPayload(model.id);
+        this.client.emit(DispatchMessage.processed, message);
       }
     }
     return model;