+++ libs/dispatch/api-models/src/lib/simulationRunReport.ts
@@ -1,62 +1,77 @@
 /**
  * @file Contains API definitions for the results of simulation run, as well as general types related to results.
  * @author Bilal Shaikh <bilalshaikh42@gmail.com>
- * 2020-12-13
- * @copyright Biosimulations Team 2020
+ * 2021-06-14
+ * @copyright Biosimulations Team 2021
  * @license MIT
  */
 
-import { ApiProperty, ApiResponseProperty } from '@nestjs/swagger';
+import {
+  ApiProperty,
+  ApiResponseProperty,
+  ApiPropertyOptional,
+} from '@nestjs/swagger';
+
 import { SchemaObject } from '@nestjs/swagger/dist/interfaces/open-api-spec.interface';
 
-export type SimulationRunReportData = SimulationRunReportDatum[];
+export type SimulationRunOutputData = SimulationRunOutputDatum[];
 
-export const SimulationRunReportDataSchema: Omit<SchemaObject, 'required'> = {
+export const SimulationRunOutputDataSchema: Omit<SchemaObject, 'required'> = {
   oneOf: [
     { type: 'array', items: { type: 'number', format: 'float' } },
     { type: 'array', items: { type: 'boolean' } },
   ],
 };
 
-export class SimulationRunReportDatum {
+export class SimulationRunOutputDatum {
   @ApiProperty({ type: String })
   public id!: string;
   @ApiProperty({ type: String })
   public label!: string;
-  @ApiProperty(SimulationRunReportDataSchema)
+  @ApiProperty({ type: String })
+  public shape!: string;
+  @ApiProperty({ type: String })
+  // "float64", "int" etc. Not the same as seddatatype
+  public type!: string;
+  @ApiProperty(SimulationRunOutputDataSchema)
   public values: (number | boolean | string)[] = [];
 }
 
-// This is used by the service when reading the results files
-export type SimulationRunReportDataStrings = { [key: string]: Array<string> };
-
-// This is used by the api to define the schema for the incoming data from the service
-export const CreateSimulationRunReportSchema: Omit<SchemaObject, 'required'> = {
-  type: 'object',
-  additionalProperties: SimulationRunReportDataSchema,
-  example: {
-    'property-1': [5.0, 2.3, 25.0],
-    'property-2': [true, true, false],
-  },
-};
-
-export class SimulationRunReport {
+export class SimulationRunOutput {
   @ApiProperty({ type: String })
   public simId!: string;
+
   @ApiProperty({ type: String })
-  public reportId!: string;
-  @ApiProperty({ type: () => [SimulationRunReportDatum] })
-  public data!: SimulationRunReportData;
+  public outputId!: string;
+
+  @ApiPropertyOptional({ type: String })
+  // "SedReport" , "SedPlot2D", "SedPlot3D" see #2442
+  public type!: string;
+
+  @ApiPropertyOptional({ type: String })
+  public name!: string;
+
+  // Dates are serialized when sending over http. Typing must be correct to prevent client from using data operations on string objects
   @ApiResponseProperty({ type: String, format: 'date-time' })
-  public created!: Date;
+  public created!: string;
+
   @ApiResponseProperty({ type: String, format: 'date-time' })
-  public updated!: Date;
+  public updated!: string;
+
+  @ApiProperty({ type: () => [SimulationRunOutputDatum] })
+  public data!: SimulationRunOutputData;
 }
 
 export class SimulationRunResults {
   @ApiResponseProperty({ type: String })
   public simId!: string;
 
-  @ApiResponseProperty({ type: () => [SimulationRunReport] })
-  public reports!: SimulationRunReport[];
+  @ApiResponseProperty({ type: String, format: 'date-time' })
+  public created!: string;
+
+  @ApiResponseProperty({ type: String, format: 'date-time' })
+  public updated!: string;
+
+  @ApiResponseProperty({ type: () => [SimulationRunOutput] })
+  public outputs!: SimulationRunOutput[];
 }